security has been added to site

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -69,9 +70,12 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required()
 def add_product(request):
-    """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access/permission to this page')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -90,8 +94,12 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required()
 def edit_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access/permission to this page')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, id=product_id)
 
     if request.method == 'POST':
@@ -117,8 +125,12 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required()
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have access/permission to this page')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, id=product_id)
 
     product.delete()
